+++ hybridse/include/sdk/base.h
@@ -129,6 +129,12 @@ class TableSet {
     virtual int32_t Size() { return 0; }
 };
 
+enum ProcedureType {
+    kUnknow = -1,
+    kReqProcedure = 0,
+    kReqDeployment,
+};
+
 class ProcedureInfo {
  public:
     ProcedureInfo() {}
@@ -140,6 +146,8 @@ class ProcedureInfo {
     virtual const hybridse::sdk::Schema& GetOutputSchema() const = 0;
     virtual const std::vector<std::string>& GetTables() const = 0;
     virtual const std::string& GetMainTable() const = 0;
+    virtual const std::string& GetMainDb() const = 0;
+    virtual ProcedureType GetType() const = 0;
 };
 
 }  // namespace sdk