+++ src/tablet/tablet_impl.cc
@@ -1512,13 +1512,14 @@ void TabletImpl::ProcessQuery(RpcController* ctrl, const openmldb::api::QueryReq
         if (request->is_debug()) {
             session.EnableDebug();
         }
+        session.SetPerformanceSensitive(request->is_performance_sensitive());
         session.SetParameterSchema(parameter_schema);
         {
             bool ok = engine_->Get(request->sql(), request->db(), session, status);
             if (!ok) {
                 response->set_msg(status.msg);
                 response->set_code(::openmldb::base::kSQLCompileError);
-                DLOG(WARNING) << "fail to compile sql " << request->sql();
+                DLOG(WARNING) << "fail to compile sql " << request->sql() << ", message: " << status.msg;
                 return;
             }
         }
@@ -3881,8 +3882,7 @@ void TabletImpl::RefreshTableInfo() {
             LOG(WARNING) << "fail to parse procedure proto. node: " << node << " value: " << value;
             continue;
         }
-        // conver to ProcedureInfoImpl
-        auto sp_info = openmldb::catalog::SchemaAdapter::ConvertProcedureInfo(sp_info_pb);
+        auto sp_info = std::make_shared<openmldb::catalog::ProcedureInfoImpl>(sp_info_pb);
         if (!sp_info) {
             LOG(WARNING) << "convert procedure info failed, sp_name: " << sp_info_pb.sp_name()
                          << " db: " << sp_info_pb.db_name();
@@ -4740,7 +4740,7 @@ void TabletImpl::CreateProcedure(RpcController* controller, const openmldb::api:
         return;
     }
 
-    auto sp_info_impl = openmldb::catalog::SchemaAdapter::ConvertProcedureInfo(sp_info);
+    auto sp_info_impl = std::make_shared<openmldb::catalog::ProcedureInfoImpl>(sp_info);
     if (!sp_info_impl) {
         response->set_msg(status.str());
         response->set_code(::openmldb::base::kCreateProcedureFailedOnTablet);